# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/images/{externalId}": {
      "get": {
        "description": "Returns binary image file by externalId. If possible, serves generated web version (webp).",
        "operationId": "ImageController_downloadByExternalId",
        "parameters": [
          {
            "name": "externalId",
            "required": true,
            "in": "path",
            "description": "Image external identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid externalId"
          },
          "404": {
            "description": "Image file not found"
          }
        },
        "summary": "Download image file by externalId",
        "tags": [
          "Images"
        ]
      }
    },
    "/api/admin/images/by-ids": {
      "post": {
        "description": "Accepts an array of internal image UUIDs and returns full metadata for each image in the same order. Deleted images are also returned if ids exist.",
        "operationId": "ImageAdminController_getImagesByIds",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetImagesByIdsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Full image metadata for all requested ids",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImageDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or one of ids has invalid UUID format"
          },
          "401": {
            "description": "Valid Bearer token is required"
          },
          "404": {
            "description": "One or more images were not found by provided ids"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get full image metadata by internal ids (admin)",
        "tags": [
          "Image Admin"
        ]
      }
    },
    "/api/admin/images/{topic}": {
      "post": {
        "description": "Uploads image to selected topic, stores image metadata, and writes outbox event {topic}.image.uploaded. Topic must be pre-registered in image topic registry.",
        "operationId": "ImageAdminController_upload",
        "parameters": [
          {
            "name": "topic",
            "required": true,
            "in": "path",
            "description": "Kafka topic/domain context for image",
            "schema": {
              "example": "catalog_product",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file",
                  "entityType",
                  "entityId"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Image file"
                  },
                  "entityType": {
                    "type": "string",
                    "description": "Domain entity type",
                    "example": "catalog.product"
                  },
                  "entityId": {
                    "type": "string",
                    "description": "Domain entity id",
                    "example": "123"
                  },
                  "imageType": {
                    "type": "string",
                    "description": "Image role/type in entity",
                    "example": "main"
                  },
                  "title": {
                    "type": "string",
                    "description": "Image title",
                    "example": "Category banner"
                  },
                  "description": {
                    "type": "string",
                    "description": "Image description",
                    "example": "Hero banner for category page"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Full image metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid topic, missing file, invalid mime type, or invalid body fields"
          },
          "401": {
            "description": "Valid Bearer token is required"
          },
          "404": {
            "description": "Topic is not configured in image topic registry"
          },
          "409": {
            "description": "Topic exists but upload event is disabled for it"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload image to topic (admin)",
        "tags": [
          "Image Admin"
        ]
      }
    },
    "/api/admin/images/{topic}/{externalId}": {
      "get": {
        "description": "Returns full image metadata in selected topic. Topic must exist in image topic registry.",
        "operationId": "ImageAdminController_getImageByExternalId",
        "parameters": [
          {
            "name": "topic",
            "required": true,
            "in": "path",
            "description": "Image topic",
            "schema": {
              "example": "catalog_product",
              "type": "string"
            }
          },
          {
            "name": "externalId",
            "required": true,
            "in": "path",
            "description": "Image external id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid topic or externalId"
          },
          "401": {
            "description": "Valid Bearer token is required"
          },
          "404": {
            "description": "Topic is not configured or image not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get image metadata by topic and externalId (admin)",
        "tags": [
          "Image Admin"
        ]
      },
      "patch": {
        "description": "Updates image metadata (imageType/title/description) and creates outbox event {topic}.image.updated. Topic must be pre-registered in image topic registry.",
        "operationId": "ImageAdminController_updateByExternalId",
        "parameters": [
          {
            "name": "topic",
            "required": true,
            "in": "path",
            "description": "Image topic",
            "schema": {
              "example": "catalog_product",
              "type": "string"
            }
          },
          {
            "name": "externalId",
            "required": true,
            "in": "path",
            "description": "Image external id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateImageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image metadata after update",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid topic, externalId, or request body"
          },
          "401": {
            "description": "Valid Bearer token is required"
          },
          "404": {
            "description": "Topic is not configured or image not found"
          },
          "409": {
            "description": "Topic exists but update event is disabled for it"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update image metadata by topic and externalId (admin)",
        "tags": [
          "Image Admin"
        ]
      },
      "delete": {
        "description": "Marks image as deleted and creates outbox event {topic}.image.deleted. Topic must be pre-registered in image topic registry.",
        "operationId": "ImageAdminController_deleteByExternalId",
        "parameters": [
          {
            "name": "topic",
            "required": true,
            "in": "path",
            "description": "Image topic",
            "schema": {
              "example": "catalog_product",
              "type": "string"
            }
          },
          {
            "name": "externalId",
            "required": true,
            "in": "path",
            "description": "Image external id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image metadata after soft-delete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid topic or externalId"
          },
          "401": {
            "description": "Valid Bearer token is required"
          },
          "404": {
            "description": "Topic is not configured or image not found"
          },
          "409": {
            "description": "Topic exists but delete event is disabled for it"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft-delete image by topic and externalId (admin)",
        "tags": [
          "Image Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Base Service API",
    "description": "REST API + GRPC endpoints",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3972"
    }
  ],
  "components": {
    "schemas": {
...
      "GetImagesByIdsDto": {
        "type": "object",
        "properties": {
          "ids": {
            "description": "Array of internal image UUIDs",
            "example": [
              "d7af3686-6d6d-4fb7-9b25-33b1ef2257c2",
              "a80d92b7-f417-4294-b85d-74c134b8e6ec"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "ids"
        ]
      },
      "ImageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Internal image record UUID"
          },
          "externalId": {
            "type": "string",
            "description": "External image identifier"
          },
          "topic": {
            "type": "string",
            "description": "Kafka topic where the image belongs",
            "example": "catalog_product"
          },
          "entityType": {
            "type": "string",
            "nullable": true,
            "example": "catalog.category",
            "description": "Domain entity type"
          },
          "entityId": {
            "type": "string",
            "nullable": true,
            "example": "42",
            "description": "Domain entity identifier"
          },
          "imageType": {
            "type": "string",
            "nullable": true,
            "example": "main",
            "description": "Image role/type for entity"
          },
          "title": {
            "type": "string",
            "nullable": true,
            "example": "Category banner",
            "description": "Image title"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Hero banner for category page",
            "description": "Image description"
          },
          "storage": {
            "type": "string",
            "description": "Storage provider",
            "enum": [
              "LOCAL",
              "S3",
              "MINIO"
            ]
          },
          "bucket": {
            "type": "string",
            "nullable": true,
            "example": null,
            "description": "Bucket name (for object storage)"
          },
          "path": {
            "type": "string",
            "description": "Physical image path"
          },
          "originalName": {
            "type": "string",
            "description": "Original uploaded filename"
          },
          "mimeType": {
            "type": "string",
            "description": "Original image MIME type"
          },
          "extension": {
            "type": "string",
            "description": "Image file extension (without dot)"
          },
          "size": {
            "type": "number",
            "description": "Image size in bytes"
          },
          "width": {
            "type": "number",
            "nullable": true,
            "example": 1280,
            "description": "Image width in pixels"
          },
          "height": {
            "type": "number",
            "nullable": true,
            "example": 720,
            "description": "Image height in pixels"
          },
          "isPublic": {
            "type": "boolean",
            "description": "Public access flag"
          },
          "checksum": {
            "type": "string",
            "nullable": true,
            "example": "d41d8cd98f00b204e9800998ecf8427e",
            "description": "File checksum (md5)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-03-13T13:57:51.388Z",
            "description": "Created timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-03-13T13:57:51.388Z",
            "description": "Updated timestamp"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": null,
            "description": "Soft-delete timestamp"
          }
        },
        "required": [
          "id",
          "externalId",
          "topic",
          "storage",
          "path",
          "originalName",
          "mimeType",
          "extension",
          "size",
          "isPublic",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateImageDto": {
        "type": "object",
        "properties": {
          "imageType": {
            "type": "string",
            "description": "New logical image type",
            "example": "gallery"
          },
          "title": {
            "type": "string",
            "description": "New image title",
            "example": "Updated category banner"
          },
          "description": {
            "type": "string",
            "description": "New image description",
            "example": "Updated hero banner for category page"
          }
        }
      },
...
    }
  }
}